Should've split these into different commits, but I don't know svn like git
1) Update svnignore to ignore README.pdf (doc output)

2) Remove README.pdf from revision control

3) Clean up complier and linker flags to useful / accurate state

4) Fix resulting warnings/errors

5) Restructure repo with nice recursive makefiles

6) Add 3 trivial testbenches
   (ironicially trivialS.s isn't working atm)



git-svn-id: svn+ssh://energy.eecs.umich.edu/svnroot/teaching/eecs373-f11/tools/sim@49 40ecb139-3c61-4c5f-bdaf-cae60a0fc81e

+++ operations/add.c
@@ -48,7 +48,7 @@ int add1(uint32_t inst) {
 
 int add2(uint32_t inst) {
 	uint8_t rd = (inst & 0x700) >> 8;
-	int32_t immed8 = (inst & 0xff);
+	uint8_t immed8 = (inst & 0xff);
 
 	uint32_t rd_val = CORE_reg_read(rd);
 	uint32_t result = rd_val + immed8;
